apps/api/pyproject.toml
[project]
name = "api"
version = "2.0.0"
description = "Piezo.AI v2 FastAPI Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.0",
    "pydantic-settings>=2.2.0",
    "structlog>=24.1.0",
    "alembic>=1.13.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.0",
    "ml-core",
    "db",
    # Phase 5: AI Research Assistant (optional)
    "langgraph>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-ollama>=0.3.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "pypdf2>=3.0.0",
    "websockets>=13.0",
]

[tool.uv]
workspace = true
